Refactor WebSocket and Simulation Services
