+++ API/protocol/api.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package proto;
+package protocol;
 
 // ============================================================================
 // CORE TYPES
